ml/requirements.txt
flask
flask-cors
tensorflow
numpy
Pillow
scikit-learn
matplotlib